Revert changes to backend streaming service.
