+++ backend/streaming/streaming/app.py
@@ -1,27 +1,16 @@
 import logging
 from aiohttp import web
 
-from .monitoring.config import Config
-from .monitoring.manager import setup_monitoring
+from .monitoring import setup_monitoring
 from .routes import setup_routes
 
 
 def create_app():
     logging.basicConfig(level=logging.INFO)
 
-    # TODO: config file/cmd line
-    config = Config(
-        kafka_bootstrap_server='kafka:29092',
-        kafka_topic_map={
-            'callListUpdate': 'callListUpdate',
-            'callInfoUpdate': 'callInfoUpdate',
-            'rosterUpdate': 'rosterUpdate'
-        }
-    )
-
     app = web.Application()
     setup_routes(app)
-    setup_monitoring(app, config)
+    setup_monitoring(app)
     return app
