+++ backend/streaming/streaming/routes.py
@@ -1,7 +1,7 @@
 import aiohttp_cors
 from aiohttp import web
 
-from .views import cancel_monitoring, get_notifications, schedule_monitoring
+from .monitoring import *
 
 
 def setup_routes(app: web.Application):
@@ -12,11 +12,5 @@ def setup_routes(app: web.Application):
             allow_headers="*",
         )
     })
-    cors.add(app.router.add_get('/notifications/{conf_name}', get_notifications))
-    cors.add(app.router.add_get('/notifications', get_notifications))
-
-    cors.add(app.router.add_post('/monitoring/{conf_name}', schedule_monitoring))
-    cors.add(app.router.add_delete('/monitoring/{conf_name}', cancel_monitoring))
-
-    cors.add(app.router.add_post('/monitoring', schedule_monitoring))
-    cors.add(app.router.add_delete('/monitoring', cancel_monitoring))
+    cors.add(app.router.add_get('/notifications/{conf_id}', get_notifications))
+    cors.add(app.router.add_post('/monitoring/{conf_id}', schedule_monitoring))
